requirements.txt
-e .
aws-cdk.aws-lambda==1.27.0
